Removed double header links
play-doc has supported header links for quite some time, but recently it
was upgraded to a version of pegdown that also added header links, using
a slightly different approach, which meant headings now had double
anchors, and two different link styles.

This turns off pegdowns header links support, so that it only uses the
old approach that we added.